backend/requirements.txt
# FastAPI and ASGI server
fastapi
uvicorn[standard]

# Docker SDK for Python
docker>=7.0.0
requests

# WebSocket support
websockets

# Google Gemini AI (latest)
google-generativeai

# Environment variables
python-dotenv

# CORS middleware
python-multipart

# Pydantic for data validation
pydantic
pydantic-settings

# Async support
aiofiles

# Rate limiting
slowapi

# Logging
loguru

# Testing (optional)
pytest
pytest-asyncio
httpx
